Split request preparation and request initiation into two separate methods
The method `request(method:requestMutation) -> Request` encapsulates
both preparing and initiating a request. When creating a `SignalProducer`
the `requestMutation` method gives a compiler error if used inside the block
for `SignalProducer` because it's denoted as `@noescape`. It will outlive the
current scope if it's captured inside the block for `SignalProducer` because we
don't know when the producer will be started. The side effects (initiation)
should happen when someone starts the `SignalProducer`, so it was split into
two methods: First prepares an `NSMutableURLRequest` with the `method` and
`requestMutation`, and the other starts a request given an `NSMutableURLRequest`