Extension safe library
This commit enables the `Allow app extension API only` option.
Because there are no non-safe apis used in the library it can be enabled without producing any warnings.
This change will the `unsafe api` warning if one uses the framework in an app extension.

SiestaUI can't be labeled `app extension safe`, because of it's NetworkActivityIndicator.